backend/pyproject.toml
[project]
name = "avatar-video-creator"
version = "0.1.0"
description = "AI educational video pipeline: script → MP3 → MP4"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    # Required by starlette.middleware.sessions (Google OAuth); not always pulled transitively.
    "itsdangerous>=2.0.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "openai>=1.55.0",
    "httpx>=0.27.0",
    "pillow>=10.4.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "boto3>=1.35.0",
    "google-auth>=2.29.0",
    "google-cloud-texttospeech>=2.16.0",
    "psycopg2-binary>=2.9.10",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
coqui = ["TTS>=0.22.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "moto[s3]>=5.0.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = ["static/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests that need ffmpeg or real binaries",
]
